requirements.txt
Twisted==15.0.0
dnspython==1.12.0
wsgiref==0.1.2
zope.interface==4.1.2
